requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
statsmodels>=0.14
prophet>=1.1
matplotlib>=3.7
seaborn>=0.13
plotly>=5.20
streamlit>=1.38
pyyaml>=6.0
